vercel-server-app/requirements.txt
Flask
gunicorn
flask-cors
python-dotenv
pymongo
python-docx
PyPDF2
paddleocr
paddlepaddle
google-generativeai
